improve cardlogin, http util, etc

+++ src/main/java/com/cdac/enrollmentstation/dto/CRSelectAppReqDto.java
@@ -15,7 +15,7 @@ import lombok.experimental.FieldDefaults;
 @AllArgsConstructor
 public class CRSelectAppReqDto {
     @JsonProperty("cardtype")
-    byte cardType;
+    int cardType;
     @JsonProperty("handle")
     int handle;
 }
